requirements.txt
openai
pymupdf
pytesseract
pillow
pandas
chromadb
sentence-transformers
streamlit
camelot-py[cv]
tabulate
